handled int max

+++ push_swap/push_swap.h
@@ -6,7 +6,7 @@
 /*   By: abeh <abeh@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/17 06:45:45 by abeh              #+#    #+#             */
-/*   Updated: 2024/07/02 15:13:28 by abeh             ###   ########.fr       */
+/*   Updated: 2024/07/03 05:48:35 by abeh             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -22,7 +22,7 @@
 
 typedef struct t_elements
 {
-	int			data;
+	long		data;
 	int			index;
 	int			bits[9];
 }				t_elements;
@@ -44,7 +44,7 @@ int				lowest_int(t_stack a);
 int				is_sorted(t_stacks *s);
 int				has_duplicates(t_stacks *s);
 int				is_valid_int(const char *str);
-int				ft_atoi(const char *c);
+long				ft_atoi(const char *c);
 
 void			index_elements(t_elements *array, int size);
 void			insertion_sort(t_elements *array, int size);
